backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
pyyaml==6.0.1

